Adding FSAE Library
This will contain all code that is reused throughout the projects. Also
this will contain headers for keeping track of which pins are used in
which project. This is so that we can drive unused pins to logic low and
keep the function that does it generic and a part of the library.
Lastly, this library will contain defines that used throughout the
project for CAN. This includes message IDs and byte locations of data.