Fixing Code and Adding FSAE Library
This is a fix for potential race conditions in the uSD source code. In
this application we don't have multiple threads, but with interrupts
present we need to be conscious of how we access volatile variables. In
the past this has been ignored but it's important that we disable
interrupts for critical sections where we can't perform atomic
operations on volatile variables.

NOTE: all this code is untested and this commit may have broken the
code.

Lastly, changes we made to make use of the new FSAE Library.