Error messages for telemetry
Added recognition of error messages in the telemetry module. Will
probably be moved to a different location, but right now it is just is a
place holder for testing. Also started the implementation of different
types of errors. Can detect (hopefully) and send error message when it
hasn't received one in a while, and can send an error message if there
is a CAN bus error.