A work in progress
So this code is not 100% working, the problems that I am having is the
implementation of the short shifts and the checking of the status
during blinks. Now after talking with Andrew, I am going to
de-implement the checking of status during blinks from the helper
methods and I will have the main method do this job. After talking with
Andrew, by doing this it will allow for greater performance. Which I am
unsure why it would but Ill discuss that with him later. In the next
day or two I should have finished code for testing.