Working Shift Lights Code
This is fully working shift lights code. I will have to ask George some
details on the documentation but other than that, it is done. Next, I
am gonna comment out all the functions and I want to shorten the code a
little bit so I will do that as well.