Complete
I believe I have everything fully implemented now. I put in the motec
and can error handling time delays as well as the can error checker in
general. I also put in a checker for oil temp and engine temp in the
motec checker. The pre-programmed simulator is difficult to change
while it is running so I want to verify my code with vector on monday.