Fix major issue with interrupt shadow register sets.
The PRISS register maps each interrupt priority level to a shadow register
set. There are 7 priority levels, and 8 shadow register sets, so any
logically thinking person would assume that level 1 is automatically
mapped to set 1, level 2 to set 2, and so on. That logically thinking
person would be incorrect. The value of the PRISS register upon reset will
be all zeros, which maps every priority level to shadow register set 0. If
left unchanged, this causes all sort of fuckery with multiple interrupts,
as they will be using the same register set without bothering to save and
restore registers. Fixing this issue cleared up just about every issue
we've been having with the PIC32 to date. I was able to remove all CLI/STI
statments as a result.

Also improved the handling of the ADC input scan trigger. Apparently all
the desired channels are included in the scan even though it's impossible
to change their trigger source due to the missing ADCTRG registers.