Fix really stupid PDM STR load bug.
Essentially, after the pulsing finished, the load could be
off (due to pulsing), so the PDM would turn it on again, thinking
it was "fully" off. This would reset the pulsing control and
another pulsing section would occur. This meant the STR load was
pulsed indefinitely, which led to a fake ~5V being seen on the
output and the starter motor failing to turn properly. Ensuring the
load is on when pulsing finishes will fix the issue.